Workaround: errors not passed through printf()
There's an issue in the C library used when working on this program
where the error values weren't propagating up through printf() so the
application can tell if printf() ran successfully or not.
e.g.:
$ ./hello > /dev/full
$ echo $?
0
$ ldd --version
ldd (Ubuntu GLIBC 2.31-0ubuntu9.16) 2.31

The work around is to not use printf() but fprintf() then fflush()
the output immediately after writing and test to see if fflush()
ran successfully.